extract init metrics into function
